server/requirements.txt
fastapi
uvicorn[standard]
transformers[torch]
torch
google-cloud-storage